From Michael Marineau:
When resolving between a group name and the local gid the
group's data structure is stored in a buffer the size of
sysconf(_SC_GETGR_R_SIZE_MAX) which is usually 1024. This
size is to small for any group larger than about 100 users.
When the limit is reached name resolution failed and the
default nogroup value is used.

The attached patch fixes this issue by attempting larger
and larger buffer sizes until the group structure fits
starting with sysconf(_SC_GETGR_R_SIZE_MAX). If the group
fits in the default size the behaviour is the same.