A while back I had a complaint that idmapper *always* gets the uid/gid mapping wrong because the default value defined in the 'Domain' field (in /etc/idmapd.conf) is never correct.
The person also noticed that by commenting out the Domain
fields caused the mappings to be correct, since DNS was used
to obtain the domain.

But then it was noticed when the domain can not be gotten from
the idmapd.conf or DNS, idmapper crashes which does seem too
robust... imho...

The first patch comments out the 'Domain' field  in idmapd.conf.
The second patch  sets the 'default_domain' variable in libnfsidmap.c to
"localdomain" when there is not a domain  defined in /etc/idmapd.conf
and a domain name can not be obtained from DNS. Of course using
"localdomain" as the domain will cause all mapping to be map to
the 'Nobody-User' user but better than having idmapper crash, imho...

So these patches do two things, allows mappings to be correct "right out of the box"
when DNS is set up correctly and stops idmapper from dying when there
is no domain name set.

Signed-off-by: Steve Dickson <steved@redhat.com>