Add numerical string translation
nfs4_owner_to_gid and nfs4_owner_to_uid will call their respective
nfs4_name_to_*id functions first.  If the name could not be resolved, we
check if the string is a numerical representation of an id number.
If it is, we convert the string to an id number.  If not, we map
the user to "nobody"

If we are unable to find a mapping for a uid or gid, we convert the
id into a numeric string to send to the server.

Signed-off-by: Bryan Schumaker <bjschuma@netapp.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Steve Dickson <steved@redhat.com>